LP #1996074: Machine details stuck at "Loading" for machines with no disks
The issue was that if you delete all disks from a machine, the testing
node's current scriptset will be empty. Then when the machine was deploy, the
creation of the installation scriptset would clean up the empty testing
scripset, leaving a reference to a deleted row.

I've changed things so that when you create a new scriptset, we only clean up
scriptsets of the same type. I also fixed the tests, since they didn't actually
create any empty sets. The reason the tests passed before was that they set the
max limit to 1.

Now the real issue is that Node.current_testing_script_set doesn't have a
foreign key in the database, and that the 'on_delete=SET_NULL' doesn't work.
But considering that might be hard to fix, I think it's worth landing this
workaround for now.