Remove the @perf_test decorator.
It's not needed, and it was somewhat broken. It reverted back to
a savepoint, unless commit_transaction was true. In that case
it simply removed the savepoint. I'm not quite sure what the intent
was there. If it was to simulate a commit, it was in the wrong
place, since it was outside the block that measures the duration.

The transaction is rolled back elsewhere for each tests, so no
need for a custom decorator. The decorator also had logic for
skipping the tests if django wasn't loaded, but I don'