Remove the custom bin/test.perf runner.
There's no reason to have a custom test runner. It's better to
use pytest directly, so that it's clear how to use it.

I also improved the way we set the random seed, so that we do
it per-test and not per-session, and print out the used values
on test failures.

I kept bin/test.perf as a symlink, in case someone is very
used to it.