Add a pytest DB fixture to set up the database for the tests.
You now no longer have to run the pytest tests that require a DB using
bin/database. Instead you run bin/pytest and it will set up the DB
automatically. It still uses the bin/database infrastructure under the hood,
and it will re-use the DB by default. So the time it takes to run the tests is
the same as before.

For the performance tests, it's possible to pass in a DB dump to use as well.
I removed pytest-django, since we want to control how the DB is
created and cleaned up.