refactor(agent)!: change Go module name
BREAKING CHANGE: Go module import path is changed

`go get` doesn't play well with Launchpad (git) [0]
We incorrectly assumed that module will be available at:
  launchpad.net/maas/maas/src/maasagent

It seems that correct way to get the module is to use this URL:
  git.launchpad.net/~maas-committers/maas.git/src/maasagent

However there will be an import error:
  module declares its path as: launchpad.net/maas/maas/src/maasagent
          but was required as: git.launchpad.net/~maas-committers/maas.git/src/maasagent

In order to solve this error, this commit changes module name to match URL
where it is available and is `go get`-able.

We introduce maas.io/core that handles ?go-get=1
query parameter and provides metadata for the `go get` utility.

Go modules can be imported as `maas.io/core/src/maasagent`

[0]: https://github.com/golang/go/blob/master/src/cmd/go/internal/vcs/vcs.go