Make tarball uploads have the correct extension.
When doing a deploy to ram, the initramfs we use look at the file
extension to determine what kind of file the image is. A name like
root-tgz won't work, it has to be root.tgz.

I migrated the filenames in the database, but not on disk. Non-ephemeral
deploys will continue working, even though the filename in the db and
on disk are different. Eventually, when the rack syncs images from
the region, the file will be renamed on disk as well.