refactor: power workflow cleanup
- Rename PowerNWorkflow to PowerManyWorkflow
During conversations we call it PowerMany, because PowerN doesn't soundwell.

- Rename Queue property to TaskQueue.
We should use Temporal terms for Temporal related code (workflows/activities). The correct term in this case is Task Queue, as queue might be something different.

- Remove custom ID for the PowerMany workflow
If we need uniqueness, then we can leave it blank and a random GUID will be generated