Support partial HTTP requests in the maas-agent proxy.
If the client requests part of the file using a Range header in the request,
the maas-agent would return a 200 response with the full file.

Now we use http.ServeContent to return the response, which knows how
to deal with the Range header. We also return partial responses as they are
from the upstream server.

This is needed, since grub sometimes requests parts of the file it
downloads.