fix(agent): custom un/marshal for the Host type
Host is being used on both Python and Go side.
In Python it is defined as:

@dataclass
class Host:
    ip: Optional[str]
    mac: Optional[str]
    hostname: Optional[str]

Which leads to certain issues with MAC address as there is no custom
UnmarshalText (like for net.IP). And since net.HardwareAddr is actually a
[]byte it will fail to Unmarshal from "ca:fe:ba:be:00:00" as it is expecting
base64-encoded string.

From https://pkg.go.dev/encoding/json:
> Array and slice values encode as JSON arrays, except that []byte encodes
> as a base64-encoded string, and a nil slice encodes as the null JSON value.