test: fix mark node failed flaky test
This test was occasionally failing because, when creating a node, we randomly
choose the status. If the status happens to be FAILED_DEPLOYMENT, the
`mark_failed` function calls `stop_workflow` and it fails because there isn't
an asyncio loop running. To solve the problem we patch the `stop_workflow`
function since we are not interested in the outcome of that function call.