test(agent): deep refactoring of resolver tests
Instead of crafting DNS messages manually use dig-like format.
It is almost instantly readable by anyone who’s dealt with DNS, and in addition
amount of lines for every test of the MAAS DNS resolver is reduced by a factor of four.

This commit also fixes incorrect behaviour for multiple questions (not tested before on the ServeDNS handler). Only a single entry in AUTHORITY and ADDITIONAL sections should be returned.