fix(agent): add rewrite rule for mmx64.efi
Some MAAS-connected DUTs specifically request mmx64.efi
via TFTP, which fails and can cause the machine
to be unbootable due to MAAS not serving this file, if
the machine was expecting to boot MokManager.

This patch is required for the MokManager EFI entry to be served by MAAS
and loaded by DUTs correctly.

Before this patch is merged, mmx64.efi will also need to be added to
the MAAS bootloaders directory

Resolves http://launchpad.net/bugs/2111335